weekten-final (1)/weekten-final/requirements.txt
bcrypt==4.2.0
streamlit
openai
python-dotenv